+++ np.Tidy/src/tidyMain.js
@@ -128,7 +128,7 @@ export async function removeDoneMarkers(params: string = ''): Promise<void> {
     }
 
     // Get num days to process from param, or by asking user if necessary
-    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays)
+    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays ?? 0)
     logDebug('removeDoneMarkers', `numDays = ${String(numDays)}`)
     // Note: can be 0 at this point, which implies process all days
 
@@ -232,7 +232,7 @@ export async function removeDoneTimeParts(params: string = ''): Promise<void> {
     }
 
     // Get num days to process from param, or by asking user if necessary
-    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays)
+    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays ?? 0)
     logDebug('removeDoneTimeParts', `numDays = ${String(numDays)}`)
     // Note: can be 0 at this point, which implies process all days
 
@@ -314,8 +314,8 @@ export async function removeDoneTimeParts(params: string = ''): Promise<void> {
 }
 
 /**
- * Remove a given section from recently-changed Notes
- * Can be passed parameters to override default time interval through an x-callback call
+ * Remove a given section (by matching on their section heading) from recently-changed Notes. Note: does not match on note title.
+ * Can be passed parameters to override default time interval through an x-callback call.
  * @author @jgclark
  * @param {?string} params optional JSON string
  */
@@ -334,7 +334,7 @@ export async function removeSectionFromRecentNotes(params: string = ''): Promise
     }
 
     // Get num days to process from param, or by asking user if necessary
-    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays)
+    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays ?? 0)
     logDebug('removeSectionFromRecentNotes', `numDays = ${String(numDays)}`)
     // Note: can be 0 at this point, which implies process all days
 
@@ -345,10 +345,11 @@ export async function removeSectionFromRecentNotes(params: string = ''): Promise
     // Decide what matching type to use
     const matchType: string = await getTagParamsFromString(params ?? '', 'matchType', config.matchType)
     logDebug('removeSectionFromRecentNotes', `matchType = ${matchType}`)
+
     // If not passed as a parameter already, ask for section heading to remove
     let sectionHeading: string = await getTagParamsFromString(params ?? '', 'sectionHeading', '')
     if (sectionHeading === '') {
-      const res: string | boolean = await getInputTrimmed("What's the heading of the section you'd like to remove from some notes?", 'OK', 'Remove Section from Notes')
+      const res: string | boolean = await getInputTrimmed(`What's the heading of the section you'd like to remove from ${numDays > 0 ? 'some' : 'all'} notes?`, 'OK', 'Remove Section from Notes')
       if (res === false) {
         return
       } else {
@@ -363,13 +364,13 @@ export async function removeSectionFromRecentNotes(params: string = ''): Promise
     // This returns all the potential matches, but some may not be headings, so now check for those
     switch (matchType) {
       case 'Exact':
-        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content === sectionHeading)
+        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content === sectionHeading && n.headingLevel !== 1)
         break
       case 'Starts with':
-        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content.startsWith(sectionHeading))
+        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content.startsWith(sectionHeading) && n.headingLevel !== 1)
         break
       case 'Contains':
-        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content.includes(sectionHeading))
+        allMatchedParas = allMatchedParas.filter((n) => n.type === 'title' && n.content.includes(sectionHeading) && n.headingLevel !== 1)
     }
     let numToRemove = allMatchedParas.length
     const allMatchedNotes = allMatchedParas.map((p) => p.note)
@@ -402,7 +403,7 @@ export async function removeSectionFromRecentNotes(params: string = ''): Promise
       if (!runSilently) {
         const res = await showMessage(`No sections with heading '${sectionHeading}' were found to remove`)
       }
-      logWarn('removeSectionFromRecentNotes', `No sections with heading '${sectionHeading}' were found to remove`)
+      logInfo('removeSectionFromRecentNotes', `No sections with heading '${sectionHeading}' were found to remove`)
     }
 
     return
@@ -444,14 +445,14 @@ export async function removeSectionFromAllNotes(params: string = ''): Promise<vo
     // If not passed as a parameter already, ask for section heading to remove
     let sectionHeading: string = await getTagParamsFromString(params ?? '', 'sectionHeading', '')
     if (sectionHeading === '') {
-      const res: string | boolean = await getInputTrimmed("What's the heading of the section you'd like to remove from all notes?", 'OK', 'Remove Section from Notes')
+      const res: string | boolean = await getInputTrimmed("What's the heading of the section you'd like to remove from ALL notes?", 'OK', 'Remove Section from Notes')
       if (res === false) {
         return
       } else {
         sectionHeading = String(res) // to help flow
       }
     }
-    logDebug('removeSectionFromRecentNotes', `sectionHeading = ${sectionHeading}`)
+    logDebug('removeSectionFromAllNotes', `sectionHeading = ${sectionHeading}`)
 
     // Ideally work out how many this will remove, and then use this code:
     // const numToRemove = 1
@@ -459,7 +460,7 @@ export async function removeSectionFromAllNotes(params: string = ''): Promise<vo
     //   if (!runSilently) {
     //     const res = await showMessageYesNo(`Are you sure you want to remove ${numToRemove} '${sectionHeading}' sections?`, ['Yes', 'No'], 'Remove Section from Notes')
     //     if (res === 'No') {
-    //       logInfo('removeSectionFromRecentNotes', `User cancelled operation`)
+    //       logInfo('removeSectionFromAllNotes', `User cancelled operation`)
     //       return
     //     }
     //   }
@@ -504,7 +505,7 @@ export async function removeTriggersFromRecentCalendarNotes(params: string = '')
     }
 
     // Get num days to process from param, or by asking user if necessary
-    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays)
+    const numDays: number = await getTagParamsFromString(params ?? '', 'numDays', config.numDays ?? 0)
     logDebug('removeTriggersFromRecentCalendarNotes', `numDays = ${String(numDays)}`)
     // Note: can be 0 at this point, which implies process all days
 
@@ -565,7 +566,7 @@ export async function logNotesChangedInInterval(params: string = ''): Promise<vo
       logDebug(pluginJson, `logNotesChangedInInterval: Starting with no params`)
     }
 
-    const numDays = config.numDays
+    const numDays = config.numDays ?? 0
     const notesList = getNotesChangedInInterval(numDays)
     const titlesList = notesList.map((m) => displayTitle(m))
     logInfo(pluginJson, `${String(titlesList.length)} Notes have changed in last ${String(numDays)} days:\n${String(titlesList)}`)
@@ -594,7 +595,7 @@ export async function removeOrphanedBlockIDs(runSilently: boolean = false): Prom
       // $FlowFixMe[incompatible-call]
       const allMatchedNotes = parasWithBlockID.map((p) => p.note)
       // logDebug('allMatchedNotes', String(allMatchedNotes.length))
-      const recentMatchedNotes = getNotesChangedInIntervalFromList(allMatchedNotes.filter(Boolean), config.numDays)
+      const recentMatchedNotes = getNotesChangedInIntervalFromList(allMatchedNotes.filter(Boolean), config.numDays ?? 0)
       const recentMatchedNoteFilenames = recentMatchedNotes.map((n) => n.filename)
       // logDebug('recentMatchedNotes', String(recentMatchedNotes.length))
       parasWithBlockID = parasWithBlockID.filter((p) => recentMatchedNoteFilenames.includes(p.note?.filename))
@@ -673,7 +674,8 @@ export async function removeOrphanedBlockIDs(runSilently: boolean = false): Prom
 }
 
 /**
- * Remove blank (or nearly blank) notes
+ * Remove blank notes
+ * Note: blank means 2 bytes or fewer (which therefore includes ones with only "# ")
  * @author @jgclark
  * @param {boolean} runSilently?
  */